+++ src/Raven.Server/Smuggler/Documents/SmugglerExporter.cs
@@ -6,6 +6,7 @@ using Raven.Server.Documents;
 using Raven.Server.Json;
 using Raven.Server.ServerWide.Context;
 using Raven.Server.Smuggler.Documents.Data;
+using Raven.Server.Smuggler.Documents.Processors;
 using Sparrow.Json;
 
 namespace Raven.Server.Smuggler.Documents
@@ -112,23 +113,17 @@ namespace Raven.Server.Smuggler.Documents
                 if (OperateOnTypes.HasFlag(DatabaseItemType.Indexes))
                 {
                     writer.WriteComma();
-                    writer.WritePropertyName(("Indexes"));
+                    writer.WritePropertyName("Indexes");
                     writer.WriteStartArray();
+                    var isFirst = true;
                     foreach (var index in _database.IndexStore.GetIndexes())
                     {
-                        if (index.Type == IndexType.Map || index.Type == IndexType.MapReduce)
-                        {
-                            var indexDefinition = index.GetIndexDefinition();
-                            writer.WriteIndexDefinition(context, indexDefinition);
-                        }
-                        else if (index.Type == IndexType.Faulty)
-                        {
-                            // TODO: Should we export them?
-                        }
-                        else
-                        {
-                            // TODO: Export auto indexes.
-                        }
+                        if (isFirst == false)
+                            writer.WriteComma();
+
+                        isFirst = false;
+
+                        IndexProcessor.Export(writer, index, context);
                     }
                     writer.WriteEndArray();
                 }
@@ -138,9 +133,15 @@ namespace Raven.Server.Smuggler.Documents
                     writer.WriteComma();
                     writer.WritePropertyName(("Transformers"));
                     writer.WriteStartArray();
+                    var isFirst = true;
                     foreach (var transformer in _database.TransformerStore.GetTransformers())
                     {
-                        writer.WriteTransformerDefinition(context, transformer.Definition);
+                        if (isFirst == false)
+                            writer.WriteComma();
+
+                        isFirst = false;
+
+                        TransformerProcessor.Export(writer, transformer, context);
                     }
                     writer.WriteEndArray();
                 }
